+++ cypress/integration/admin/professions/edit.spec.ts
@@ -215,15 +215,13 @@ describe('Editing an existing profession', () => {
       cy.translate('app.continue').then((buttonText) => {
         cy.get('button').contains(buttonText).click();
       });
-      cy.checkIndexedSummaryListRowValue(
+      cy.checkSummaryListRowValue(
         'professions.form.label.legislation.nationalLegislation',
         'Updated legislation',
-        1,
       );
-      cy.checkIndexedSummaryListRowValue(
-        'professions.form.label.legislation.nationalLegislation',
+      cy.checkSummaryListRowValue(
+        'professions.form.label.legislation.optionalNationalLegislation',
         'Second legislation',
-        2,
       );
       cy.checkIndexedSummaryListRowValue(
         'professions.form.label.legislation.link',
@@ -464,7 +462,7 @@ describe('Editing an existing profession', () => {
         cy.checkAccessibility();
 
         cy.checkSummaryListRowValue(
-          'professions.form.label.qualifications.moreInformationUrl',
+          'professions.show.qualification.moreInformationUrl',
           'http://example.com/more-info',
         );
       });